drivers: gpio: remove not required header
Move the content of the header for the MCP23S17
into the source file.
